refactor: move Wheel trait and implementation to a new file
